refactor: move content row details getting to the content row component
