+++ components/contentRow.js
@@ -1,7 +1,43 @@
+import firebaseApp from "../firebase";
+import {
+  getFirestore,
+  collection,
+  query,
+  where,
+  limit,
+  getDocs,
+  documentId,
+} from "firebase/firestore";
+import { useState, useEffect } from "react";
 import ContentCard from "./contentCard";
 
-export default function ContentRow({ content, authors, children }) {
-  console.log(authors);
+export default function ContentRow({ content, children }) {
+  const [authors, setAuthors] = useState([]);
+  const db = getFirestore(firebaseApp);
+  const fetchContentRowDetails = async () => {
+    if (content && content.length > 0) {
+      const authorsReferences = content.map((doc) => {
+        return doc.data.owner.id;
+      });
+
+      const authorsQuery = query(
+        collection(db, "users"),
+        where(documentId(), "in", authorsReferences)
+      );
+
+      const authorsData = await getDocs(authorsQuery);
+      const authorsFromDb = authorsData.docs.map((doc) => {
+        return {
+          id: doc.id,
+          data: doc.data(),
+        };
+      });
+      setAuthors(authorsFromDb);
+    }
+  };
+  useEffect(() => {
+    fetchContentRowDetails();
+  }, [content]);
   return (
     <>
       <div className="bg-white shadow-md my-4 mx-2 p-8 rounded w-full">
@@ -14,9 +50,13 @@ export default function ContentRow({ content, authors, children }) {
                 key={eachContent.id}
                 content={eachContent.data}
                 author={
-                  authors.filter((author) => {
-                    return author.id == eachContent.data.owner.id;
-                  })[0].data
+                  authors
+                    .filter((author) => {
+                      return author.id == eachContent.data.owner.id;
+                    })
+                    .map((author) => {
+                      return author.data;
+                    })[0]
                 }
               />
             ))}
